refactor(checks): move check parsing/discovery to goose::checks
Address @aaif-goose review feedback by lifting Check parsing and
discovery out of `goose-cli` into a new `goose::checks` module that
sits alongside `sources` / `skills`.

- Add `SourceType::Check` to the SDK so checks can flow through the
  same listing/UI pipeline as skills and projects.
- Create `crates/goose/src/checks/mod.rs` containing `Check`,
  `DiscoveredReview`, `discover`, `global_checks_dirs`, and
  `candidate_scope_dirs`. Reuses `sources::parse_frontmatter` so
  checks share the YAML pipeline with skills and projects.
- Wire `SourceType::Check` into `sources::list_sources` via a new
  `Check::to_source_entry` adapter that stuffs per-check tunables
  (`model`, `turn-limit`, `tools`, `severity-default`,
  `scope_dir`) into `SourceEntry.properties`.
- Delete `goose-cli/src/commands/review/{check,discover}.rs` and
  point the CLI's review pipeline (`handler`, `orchestrator`,
  `prompt`) at `goose::checks` instead.

Net: -719 / +33 in goose-cli plus the new module in goose. All 19
`checks` tests and 37 `commands::review` tests pass; `cargo fmt` /
`cargo clippy --all-targets -- -D warnings` clean.

Tracking: AGNTOPS-23
Amp-Thread-ID: https://ampcode.com/threads/T-019e282e-50fe-7072-9130-24ddbb6bb852
