+++ crates/goose-sdk/src/custom_requests.rs
@@ -818,6 +818,7 @@ pub enum SourceType {
     Subrecipe,
     Agent,
     Project,
+    Check,
 }
 
 impl std::fmt::Display for SourceType {
@@ -829,6 +830,7 @@ impl std::fmt::Display for SourceType {
             SourceType::Subrecipe => write!(f, "subrecipe"),
             SourceType::Agent => write!(f, "agent"),
             SourceType::Project => write!(f, "project"),
+            SourceType::Check => write!(f, "check"),
         }
     }
 }
