+++ crates/goose/src/sources.rs
@@ -47,6 +47,7 @@ fn require_listable_type(source_type: Option<SourceType>) -> Result<SourceType, 
         SourceType::BuiltinSkill => Ok(SourceType::BuiltinSkill),
         SourceType::Project => Ok(SourceType::Project),
         SourceType::Agent => Ok(SourceType::Agent),
+        SourceType::Check => Ok(SourceType::Check),
         other => Err(Error::invalid_params().data(format!(
             "Source type '{}' is not supported for listing.",
             other
@@ -937,6 +938,26 @@ pub fn list_sources_with_roots(
             }
             SourceType::Agent => {
                 sources.extend(list_agent_sources(project_dir, additional_roots));
+            }
+            SourceType::Check => {
+                let working_dir = project_dir
+                    .map(str::trim)
+                    .filter(|p| !p.is_empty())
+                    .map(PathBuf::from);
+                let discovered = match working_dir.as_deref() {
+                    Some(root) => crate::checks::discover(root, &[])
+                        .map_err(|e| Error::internal_error().data(e.to_string()))?,
+                    None => crate::checks::DiscoveredReview::default(),
+                };
+                for check in discovered.checks {
+                    let global = check.path.starts_with(
+                        crate::checks::global_checks_dirs()
+                            .first()
+                            .map(PathBuf::as_path)
+                            .unwrap_or_else(|| Path::new("")),
+                    );
+                    sources.push(check.to_source_entry(global));
+                }
             }
             SourceType::Recipe | SourceType::Subrecipe => {
                 return Err(Error::invalid_params()
